Refactor: Extract DOSBox path handling logic into `DosPathResolver`
Moved DOSBox path parsing, sanitization, and relative-to-absolute path resolution methods to a new `DosPathResolver` module. Updated `ScriptEditService` to utilize these functions, improving modularity and maintaining cleaner code separation. Adjusted CMake and build structure accordingly.
